Overhaul jacket copying
This changes the name format that the jackets must be stored in. See
the Readme for more details. It also makes a minor change to support
threading in that it checks the source directory instead of the output
directory for easier difficulties. The root of this issue was that a
thread could complete processing a difficulty before its easier
diffs, thus preventing it from finding the jacket of the easier diff.