Remove pointless code
When making the `time_iter` method it seems that I completely forgot
how iterators worked and so added those pointless method reference
parameters. Since the caller gets information on the timing and events
by definition of an iterator, there is no reason to be calling checks
that the caller could just do themselves.